requirements.txt
autoreject==0.2.2
dash==2.0.0
dash_bootstrap_components==1.0.0
Flask==2.1.2
joblib==1.1.0
jupyter_dash==0.4.0
matplotlib==3.4.3
mne==0.23.4
numpy==1.21.2
pandas==1.4.2
Pillow==9.2.0
plotly==5.3.1
scipy==1.7.1
scikit-image==0.19.2
scikit-learn==1.0.2
torch==1.11.0
torchvision==0.12.0
werkzeug==2.0.0